+++ src/Symfony/Component/DependencyInjection/Dumper/YamlDumper.php
@@ -18,7 +18,6 @@ use Symfony\Component\DependencyInjection\Argument\ArgumentInterface;
 use Symfony\Component\DependencyInjection\Argument\ClosureProxyArgument;
 use Symfony\Component\DependencyInjection\Argument\IteratorArgument;
 use Symfony\Component\DependencyInjection\Argument\ServiceClosureArgument;
-use Symfony\Component\DependencyInjection\Argument\ServiceLocatorArgument;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Parameter;
@@ -263,8 +262,6 @@ class YamlDumper extends Dumper
                 $tag = 'iterator';
             } elseif ($value instanceof ClosureProxyArgument) {
                 $tag = 'closure_proxy';
-            } elseif ($value instanceof ServiceLocatorArgument) {
-                $tag = 'service_locator';
             } else {
                 throw new RuntimeException(sprintf('Unspecified Yaml tag for type "%s".', get_class($value)));
             }