+++ src/Symfony/Component/DependencyInjection/Loader/YamlFileLoader.php
@@ -15,7 +15,6 @@ use Symfony\Component\DependencyInjection\Alias;
 use Symfony\Component\DependencyInjection\Argument\ArgumentInterface;
 use Symfony\Component\DependencyInjection\Argument\ClosureProxyArgument;
 use Symfony\Component\DependencyInjection\Argument\IteratorArgument;
-use Symfony\Component\DependencyInjection\Argument\ServiceLocatorArgument;
 use Symfony\Component\DependencyInjection\ChildDefinition;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\DependencyInjection\Definition;
@@ -670,19 +669,6 @@ class YamlFileLoader extends FileLoader
                 }
 
                 return new IteratorArgument($this->resolveServices($argument, $file, $isParameter));
-            }
-            if ('service_locator' === $value->getTag()) {
-                if (!is_array($argument)) {
-                    throw new InvalidArgumentException('"!service_locator" tag only accepts mappings.');
-                }
-
-                foreach ($argument as $v) {
-                    if (!is_string($v) || 0 !== strpos($v[0], '@') || 0 === strpos($v[0], '@@')) {
-                        throw new InvalidArgumentException('"!service_locator" tagged values must be {key: @service} mappings.');
-                    }
-                }
-
-                return new ServiceLocatorArgument($this->resolveServices($argument, $file, $isParameter));
             }
             if ('closure_proxy' === $value->getTag()) {
                 if (!is_array($argument) || array(0, 1) !== array_keys($argument) || !is_string($argument[0]) || !is_string($argument[1]) || 0 !== strpos($argument[0], '@') || 0 === strpos($argument[0], '@@')) {