+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/php/services1-1.php
@@ -8,7 +8,6 @@ use Symfony\Component\DependencyInjection\Exception\InvalidArgumentException;
 use Symfony\Component\DependencyInjection\Exception\LogicException;
 use Symfony\Component\DependencyInjection\Exception\RuntimeException;
 use Symfony\Component\DependencyInjection\ParameterBag\FrozenParameterBag;
-use Symfony\Component\DependencyInjection\ServiceLocator;
 
 /**
  * Container.