+++ src/Symfony/Component/DependencyInjection/CHANGELOG.md
@@ -4,10 +4,10 @@ CHANGELOG
 3.3.0
 -----
 
+ * added "container.service_locator" tag for defining service-locator services
  * added anonymous services support in YAML configuration files using the `!service` tag.
- * [EXPERIMENTAL] added "TypedReference" and "ServiceClosureArgument" for creating service-locator services
+ * added "TypedReference" and "ServiceClosureArgument" for creating service-locator services
  * [EXPERIMENTAL] added "instanceof" section for local interface-defined configs
- * added "service-locator" argument for lazy loading a set of identified values and services
  * [EXPERIMENTAL] added prototype services for PSR4-based discovery and registration
  * added `ContainerBuilder::getReflectionClass()` for retrieving and tracking reflection class info
  * deprecated `ContainerBuilder::getClassResource()`, use `ContainerBuilder::getReflectionClass()` or `ContainerBuilder::addObjectResource()` instead